# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DOT to DOCX Converter API
  description: |
    High-performance document conversion service for converting legacy Word template files (.dot) to modern Word format (.docx).

    This service provides a robust, scalable solution for batch converting DOT files using LibreOffice as the conversion engine.

    ## Features
    - Single file and batch conversion support
    - Asynchronous processing with job queue
    - Redis-backed job management
    - Auto-scaling based on load
    - Health monitoring endpoints
    - Prometheus metrics

    ## Usage
    1. Upload a .dot file using POST /api/v1/convert
    2. Monitor job status using GET /api/v1/jobs/{id}
    3. Download the converted file when complete

  version: 1.0.0
  contact:
    name: Alterspective Engine
    url: https://github.com/Alterspective-Engine/dot-to-docx-converter
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://dot-to-docx-converter-prod.lemondesert-9ded9ffc.eastus.azurecontainerapps.io
    description: Production server
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /api/v1/convert:
    post:
      summary: Convert single DOT file to DOCX (Asynchronous)
      description: Upload a .dot file for conversion to .docx format. The conversion is processed asynchronously in a queue. Use this for large files or batch processing.
      tags:
        - Conversion
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: The .dot file to convert
                priority:
                  type: integer
                  default: 0
                  description: Job priority (higher values processed first)
                metadata:
                  type: object
                  additionalProperties:
                    type: string
                  description: Optional metadata to attach to the job
      responses:
        '202':
          description: Job accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          description: Invalid request (file missing or wrong format)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/batch:
    post:
      summary: Batch convert multiple DOT files
      description: Submit multiple files for batch conversion. Useful for processing entire directories.
      tags:
        - Conversion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchConvertRequest'
      responses:
        '202':
          description: Batch job accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  batch_id:
                    type: string
                    format: uuid
                    description: Unique batch identifier
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/JobResponse'
                  count:
                    type: integer
                    description: Number of jobs created
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/jobs/{id}:
    get:
      summary: Get job status
      description: Retrieve the current status and details of a conversion job
      tags:
        - Jobs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Job ID
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/jobs/{id}/cancel:
    post:
      summary: Cancel a job
      description: Cancel a pending conversion job
      tags:
        - Jobs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Job ID
      responses:
        '200':
          description: Job cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Job not found
        '400':
          description: Job cannot be cancelled (already processing or completed)
  /api/v1/jobs:
    get:
      summary: List jobs
      description: List all jobs with optional status filtering
      tags:
        - Jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
          description: Filter by job status
      responses:
        '200':
          description: Job list
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/JobResponse'
                  count:
                    type: integer
  /api/v1/download/{id}:
    get:
      summary: Download converted file
      description: Download the converted DOCX file for a completed job
      tags:
        - Jobs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Job ID
      responses:
        '200':
          description: Converted file
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
        '404':
          description: File not found or job not completed
...
components:
  schemas:
    JobResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
          description: Unique job identifier
        status:
          type: string
          enum: [pending, processing, completed, failed]
          description: Current job status
        input_path:
          type: string
          description: Path to the input file
        output_path:
          type: string
          description: Path to the output file
        created_at:
          type: string
          format: date-time
          description: Job creation timestamp
        started_at:
          type: string
          format: date-time
          description: Processing start time
        completed_at:
          type: string
          format: date-time
          description: Processing completion time
        duration:
          type: string
          description: Processing duration
        error:
          type: string
          description: Error message if job failed
        download_url:
          type: string
          description: URL to download the converted file
    BatchConvertRequest:
      type: object
      required:
        - source
        - destination
        - files
      properties:
        source:
          type: string
          description: Source directory path
        destination:
          type: string
          description: Destination directory path
        files:
          type: array
          items:
            type: string
          description: List of file paths to convert
          maxItems: 1000
        priority:
          type: integer
          default: 0
          description: Job priority
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message